# backend/KokoSkyAPIDOC/kokosky-api-doc-v1.0.0.yml
openapi: 3.0.3
info:
  title: Weather Forecast APIs, KokoSky, Inc.
  version: 1.0.0
  description: |-
    This is the design documentation for KokoSky weather API
  termsOfService: http://kokosky.net/terms
  contact:
    email: contact@kokosky.net
    name: KokoSky, Inc.
    url: https://skyweather.net/contacts
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
servers:
  - url: http://localhost:{port}
    variables:
      port:
        default: "8080"
    description: This is for development testing
...
paths:
...
  /api/v1/locations{code}:
    summary: Operations for managing a specific location identified by location code
    parameters:
      - $ref: '#/components/parameters/LocationCode'
    get:
      summary: Returns details of a location
      description: Clients use this API to find an existing location in the database by a specific code
      operationId: getLocation
      tags:
        - Location
      responses:
        '200':
          description: >
            Successful retrieved a specific location based on code.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '404':
          description: No location found with the given code.
    delete:
      summary: Removes an existing location found by given code.
      description: Clients use this API to remove existing location in the database by a specific code
      operationId: deleteLocation
      tags:
        - Location
      responses:
        '204':
          description: Successful removed location.
        '404':
          description: No location found with the given code.
...
  /api/v1/daily:
    summary: >-
      Operation for accessing daily weather forecast information based on
      client's IP address
    description: Clients use this API to get forecast information about weather in the upcoming days
    get:
      summary: >-
        Returns daily weather forecast information for the location based on client's IP address
      description: >
        Clients use this API to get forecast about weather in the upcoming days. Location is determined automatically based on client's IP address.
      operationId: getDailyForecastByIPAddress
      tags:
        - Daily Forecast
      responses:
        '200':
          description: >-
            Successful retrieval operation. A JSON object representing daily forecast
            data available for the client's location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyWeather'
              examples:
                dailyWeatherExample:
                  $ref: '#/components/examples/dailyWeatherExample'
        '204':
          description: No daily forecast data available
        '400':
          description: bad request. Could not determine client's IP address
        '404':
          description: No managed location found for the client's IP address
  /api/v1/daily/{code}:
    summary: >-
      Operations for accessing and updating daily weather forecast data for a specific location
    description: >-
      Clients use these APIs to get and modify daily weather forecast information based on location code
    parameters:
      - $ref: '#/components/parameters/LocationCode'
    get:
      summary: Returns daily weather forecast information for a specific location code
      description: >-
        Clients use this API to get forecast about weather in the upcoming
        days, for the given location code
      operationId: getDailyForecastByLocationCode
      tags:
        - Daily Forecast
      responses:
        '200':
          description: >-
            successful retrieval operation. A JSON object representing daily forecast data available for the given location code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyWeather'
              examples:
                dailyWeatherExample:
                  $ref: '#/components/examples/dailyWeatherExample'
        '204':
          description: No daily forecast data available
        '404':
          description: No managed location found for the given location code
    put:
      summary: >-
        Update daily weather forecast information for a location specified by the given location code.
      description: >-
        Clients use this API to update data about weather forecast
        information in the upcoming days, based on given location code. Successful update operation will replace all existing data.
      operationId: updateDailyForecast
      tags:
        - Daily Forecast
      requestBody:
        description: >-
          a JSON object that represents the daily weather forecast data that needs to be updated. It must be an array of daily forecast information as shown in the below example.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailyWeather'
            example:
              [
                {
                  "day_of_month": 10,
                  "month": 2,
                  "min_temp": 6,
                  "max_temp": 12,
                  "precipitation": 60,
                  "status": "Cloudy"
                },
                {
                  "day_of_month": 11,
                  "month": 2,
                  "min_temp": 7,
                  "max_temp": 14,
                  "precipitation": 70,
                  "status": "Sunny"
                },
                {
                  "day_of_month": 12,
                  "month": 2,
                  "min_temp": 6,
                  "max_temp": 10,
                  "precipitation": 90,
                  "status": "Rainy"
                }
              ]
      responses:
        '200':
          description: Successfully updated daily weather forecast - all previous data is replaced.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailyWeather'
              examples:
                dailyWeatherExample:
                  $ref: '#/components/examples/dailyWeatherExample'
        '400':
          description: >-
            Bad request. Request body contains empty array (no data) or there are some invalid field values in daily forecast information.
        '404':
          description: No managed location found for the given code
  /api/v1/full:
    summary: Operation for accessing full weather information based on client's IP address
    description: >-
      Clients use this API get weather information of a location, including realtime data, hourly and daily forecast, with location identified based on client's IP address
    get:
      summary: Returns full weather forecast information based on client's IP address
      description: >-
        Clients use this API to get full weather information about weather of a location determined based on client's IP address
      operationId: getFullWeatherByIPAddress
      tags:
        - Full Weather Forecast
      responses:
        '200':
          description: >-
            successful retrieval operation. A JSON object represents full weather forecast information, which is aggregation of realtime weather, hourly forecast and daily forecast
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullWeather'
              examples:
                fullWeatherExample1:
                  $ref: '#/components/examples/fullWeatherExample1'
                fullWeatherExample2:
                  $ref: '#/components/examples/fullWeatherExample2'
                fullWeatherExample3:
                  $ref: '#/components/examples/fullWeatherExample3'
                fullWeatherExample4:
                  $ref: '#/components/examples/fullWeatherExample4'
        '400':
          description: Bad request. Could not determine client's IP address
        '404':
          description: No managed location found for the client's IP address
    put:
      summary: >-
        Updates full weather forecast information of the location specified by the given code
      description: >-
        Clients use this API to update realtime, hourly and daily forecast
        of a location at once (full weather information)
      operationId: updateFullWeather
      tags:
        - Full Weather Forecast
      requestBody:
        description: >-
          a JSON object that represents the full weather forecast information that
          needs to be updated. Realtime weather, hourly forecast and daily forecast must be present. Location information is not required.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FullWeather'
            examples:
              updateFullWeatherExample:
                $ref: '#/components/examples/updateFullWeatherExample'
      responses:
        '200':
          description: successful update operation. Full weather forecast data updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullWeather'
              examples:
                fullWeatherExample1:
                  $ref: '#/components/examples/fullWeatherExample1'
        '400':
          description: >-
            Bad request. Request body contains some invalid field values in realtime weather, hourly forecast or daily forecast
        '404':
          description: No managed location found for the given code
components:
  schemas:
    Location:
      type: object
      properties:
        code:
          type: string
          description: unique code of a location
          example: LACA_US
        city_name:
          type: string
          description: name of the city
          example: Los Angeles
        region_name:
          type: string
          description: name of the region, state or province
          example: California
        country_code:
          type: string
          description: 2 characters representing country code based on ISO 3166 standard
          example: US
        country_name:
          type: string
          description: name of the country based on ISO 3166 standard
          example: United States of America
        enabled:
          type: boolean
          description: whether the location enabled (true) or disabled (false)
          example: true
...
    DailyWeather:
      type: object
      properties:
        location:
          type: string
          description: name of location, in form of "city name, state/region name, country name"
          example: New York, New York, US
        daily_forecast:
          type: array
          items:
            type: object
            properties:
              day_of_month:
                type: number
                description: the forecasted day
                example: 8
              month:
                type: number
                description: the forecasted month (1-12)
                example: 8
              min_temp:
                type: number
                format: int32
                description: the forecasted minimum temperature (default in Celsius degree)
                example: 12
              max_temp:
                type: number
                format: int32
                description: the forecasted maximum temperature (default in Celsius degree)
                example: 12
              precipitation:
                type: number
                format: int32
                description: percentage chance of rain
                example: 67
              status:
                type: string
                description: current weather status
                example: Cloudy
    FullWeather:
      type: object
      properties:
        location:
          type: string
          description: name of location, in form of "city name, state/region name, country name"
          example: New York, New York, US
        realtime:
          type: object
          properties:
            temperature:
              type: number
              format: int32
              description: current temperature of the location, default in Celcius degree
              example: 12
            humidity:
              type: number
              format: int32
              description: percentage of humidity
              example: 72
            precipitation:
              type: number
              format: int32
              description: percentage chance of rain
              example: 50
            status:
              type: string
              description: current weather status
              example: Sunny
            wind_speed:
              type: number
              format: int32
              description: current wind speed (default unit is km/h)
              example: 8
            last_updated:
              type: string
              format: date-time
              description: the last time weather data is updated
  parameters:
    LocationCode:
      name: code
      in: path
      description: code of location
      required: true
      schema:
        type: string
        example: NYC_US
...
  examples:
...
    dailyWeatherExample:
      summary: daily weather forecast information
      value: |-
        {
          "location": "San Francisco, California, United States of America",
          "daily_forecast":
          [
            {
              "day_of_month": 10,
              "month": 2,
              "min_temp": 6,
              "max_temp": 12,
              "precipitation": 60,
              "status": "Cloudy"
            },     
            {
              "day_of_month": 11,
              "month": 2,
              "min_temp": 7,
              "max_temp": 14,
              "precipitation": 70,
              "status": "Sunny"
            },   
            {
              "day_of_month": 12,
              "month": 2,
              "min_temp": 6,
              "max_temp": 10,
              "precipitation": 90,
              "status": "Rainy"
            }          
          ]
        
        }
    fullWeatherExample1:
      summary: full weather forecast information
      value: |-
        {
          "location": "San Francisco, California, United States of America",
          "realtime_weather": {
            "temperature": 12,
            "humidity": 32,
            "precipitation": 88,
            "status": "Cloudy",
            "wind_speed": 5,
            "last_updated": "2023-02-02 19:23:00"
          },      
          "hourly_forecast":
          [
            {
              "hour_of_day": 8,            
              "temperature": 12,
              "precipitation": 88,
              "status": "Cloudy"
            },
            {
              "hour_of_day": 9,
              "temperature": 13,
              "precipitation": 86,
              "status": "Cloudy"
            },
            {
              "hour_of_day": 10,
              "temperature": 15,
              "precipitation": 60,
              "status": "Sunny"
            }                
          ],          
          "daily_forecast":
          [
            {
              "day_of_month": 10,
              "month": 2,
              "min_temp": 6,
              "max_temp": 12,
              "precipitation": 60,
              "status": "Cloudy"
            },     
            {
              "day_of_month": 11,
              "month": 2,
              "min_temp": 7,
              "max_temp": 14,
              "precipitation": 70,
              "status": "Sunny"
            },   
            {
              "day_of_month": 12,
              "month": 2,
              "min_temp": 6,
              "max_temp": 10,
              "precipitation": 90,
              "status": "Rainy"
            }          
          ]
        
        }
    fullWeatherExample2:
      summary: realtime weather not present
      value: |-
        {
          "location": "San Francisco, California, United States of America",    
          "hourly_forecast":
          [
            {
              "hour_of_day": 8,            
              "temperature": 12,
              "precipitation": 88,
              "status": "Cloudy"
            },
            {
              "hour_of_day": 9,
              "temperature": 13,
              "precipitation": 86,
              "status": "Cloudy"
            }            
          ],          
          "daily_forecast":
          [
            {
              "day_of_month": 10,
              "month": 2,
              "min_temp": 6,
              "max_temp": 12,
              "precipitation": 60,
              "status": "Cloudy"
            },     
            {
              "day_of_month": 11,
              "month": 2,
              "min_temp": 7,
              "max_temp": 14,
              "precipitation": 70,
              "status": "Sunny"
            }        
          ]
        
        }
    fullWeatherExample3:
      summary: hourly forecast and daily forecast not available
      value: |-
        {
          "location": "San Francisco, California, United States of America",   
          "realtime_weather": {
            "temperature": 12,
            "humidity": 32,
            "precipitation": 88,
            "status": "Cloudy",
            "wind_speed": 5,
            "last_updated": "2023-02-02 19:23:00"
          },            
          "hourly_forecast":[ ],        
          "daily_forecast": [ ]
        }
    fullWeatherExample4:
      summary: no realtime weather, no hourly forecast, no daily forecast
      value: |-
        {
          "location": "San Francisco, California, United States of America", 
          "hourly_forecast":[ ],        
          "daily_forecast": [ ]
        }
    updateFullWeatherExample:
      summary: full weather forecast information to be updated
      value: |-
        {
          "realtime_weather": {
            "temperature": 12,
            "humidity": 32,
            "precipitation": 88,
            "status": "Cloudy",
            "wind_speed": 5,
            "last_updated": "2023-02-02 19:23:00"
          },      
          "hourly_forecast":
          [
            {
              "hour_of_day": 8,            
              "temperature": 12,
              "precipitation": 88,
              "status": "Cloudy"
            },
            {
              "hour_of_day": 9,
              "temperature": 13,
              "precipitation": 86,
              "status": "Cloudy"
            },
            {
              "hour_of_day": 10,
              "temperature": 15,
              "precipitation": 60,
              "status": "Sunny"
            }                
          ],          
          "daily_forecast":
          [
            {
              "day_of_month": 10,
              "month": 2,
              "min_temp": 6,
              "max_temp": 12,
              "precipitation": 60,
              "status": "Cloudy"
            },     
            {
              "day_of_month": 11,
              "month": 2,
              "min_temp": 7,
              "max_temp": 14,
              "precipitation": 70,
              "status": "Sunny"
            },   
            {
              "day_of_month": 12,
              "month": 2,
              "min_temp": 6,
              "max_temp": 10,
              "precipitation": 90,
              "status": "Rainy"
            }          
          ]
        
        }